Compatibility with older go version
Before 1.7, SetIndent() was not available for the JSON encoder. To
enable compatability with older versions of go, we simply skip the
JSON indenting in 1.6.x and below.

This is probably temporary until 1.6 use dies out.

Note that the unit tests do not pass on go versions before 1.7 because
they expect the JSON to be nicely indented.  I don't intend to fix this,
but patches welcome if anyone else cares enough.